\name{countpeaks}
\alias{countpeaks}
\title{countpeaks}
\description{Count the number of signals in the intensity vector.}
\usage{
countpeaks(intensity_vec, missing_val=0)
}
\arguments{
  \item{intensity_vec}{Vector of intensities}
  \item{missing_val}{How are the missing values represented
  in the intensity vector? eg: NA or 0}
}
\value{Number of peaks with signals.
}
