\name{xMSanalyzer-package}
\alias{xMSanalyzer-package}
\alias{xMSanalyzer}
\docType{package}
\title{xMSanalyzer}
\description{Set of utilities for improved feature detection, 
quality assessment, overlap analysis, and batch annotation 
of metabolomics data.
}
\details{
\tabular{ll}{
Package: \tab xMSanalyzer\cr
Type: \tab Package\cr
Version: \tab 2.0.6.1\cr
Date: \tab 2015-5-21\cr
License: \tab gpl2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Karan Uppal
Maintainer: <kuppal2@emory.edu>
}
\references{
Literature or other references for background information
Uppal et. al. xMSanalyzer: automated pipeline for improved feature detection and downstream analysis
of large-scale, non-targeted metabolomics data, BMC Bioinformatics, 2013
Tianwei Yu. apLCMS - Adaptive Processing of High-Resolution LC/MS Data. Bioinformatics. 2009. 
C.A. Smith, E.J. Want, G.C. Tong, R. Abagyan, and G. Siuzdak. XCMS: Processing Mass Spectrometry Data for
Metabolite Profiling Using Nonlinear Peak Alignment, Matching, and Identification. Analytical Chemistry, 2006.
Tautenhahn R, Böttcher C, Neumann S. Highly sensitive feature detection for high resolution LC/MS. 
BMC Bioinformatics. 2008 Nov 28.
}
\keyword{ xMSanalyzer }


