\name{xMSwrapper}
\alias{xMSwrapper}
\title{xMSwrapper}
\description{
Tells user about the usage of wrapper functions for apLCMS and XCMS. 
}
\usage{xMSwrapper()
}
\references{put references to the literature/web site here
}
\author{Karan Uppal <kuppal2@emory.edu>}
\note{
Shows the usage for using apLCMS and XCMS wrapper functions
}
\seealso{xMSwrapper.apLCMS, xMSwrapper.XCMS.centWave, 
xMSwrapper.XCMS.matchedFilter}
\keyword{ ~xMSwrapper }
\keyword{ ~apLCMS }
\keyword{ ~XCMS }